+++ README.md
@@ -38,3 +38,24 @@
 - [x] 모든 플레이어가 카드를 더 받지 않는다고 결정하면 게임이 종료된다.
   - 총 점수가 21이 넘으면 더이상 카드를 받지 못하고, 이는 더 받지 않는 걸로 판단한다.
 - [x] 게임이 종료되면 받은 카드와 게임 결과를 출력한다.
+
+
+## 🚀 3단계 - 블랙잭(딜러)
+### 추가 기능 요구사항
+- 딜러는 처음에 받은 2장의 합계가 16이하이면 반드시 1장의 카드를 추가로 받아야 하고, 17점 이상이면 추가로 받을 수 없다.
+- 딜러가 21을 초과하면 그 시점까지 남아 있던 플레이어들은 가지고 있는 패에 상관 없이 승리한다.
+- 게임을 완료한 후 각 플레이어별로 승패를 출력한다.
+
+### 추가 프로그래밍 요구사항
+- 3개 이상의 인스턴스 변수를 가진 클래스를 쓰지 않는다.
+- 딜러와 플레이어에서 발생하는 중복 코드를 제거해야 한다.
+
+### TODO
+- [ ] 딜러는 처음에 플레이어와 똑같이 2장을 받는다.
+  - Player를 상위 객체로 하고,
+    - `draw()`, `canDraw()`를 추상 메서드로 선언하고, 하위 객채에서 구현
+  - Dealer는 Player를 상속하고, 딜러 역할을 한다.
+  - 참가자(Participant)는 Player를 상속하고, 게임 참가자 역할을 한다.
+- [ ] 플레이어가 카드를 모두 받은 후에는, 딜러가 처음에 받은 카드 2장의 합계가 16이하면 반드시 1장의 카드를 추가로 받아야한다.
+- [ ] 딜러가 추가로 받은 후(받을 수 있는 경우에), 경기는 종료되고 승패가 결정된다.
+- [ ] 위 과정을 화면에 출력한다.
