+++ src/main/kotlin/blackjack/domain/card/CardShape.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class CardShape(val korean: String) {
     HEART("하트"), CLUB("클로버"), DIAMOND("다이아몬드"), SPADE("스페이드")
