+++ src/main/kotlin/blackjack/domain/card/Deck.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 class Deck {
