+++ src/main/kotlin/blackjack/domain/card/DrewCards.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 import blackjack.domain.rule.ScoringRule
