+++ src/main/kotlin/blackjack/domain/rule/ScoringRule.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.rule
 
-import blackjack.domain.Card
+import blackjack.domain.card.Card
 
 interface ScoringRule {
