+++ src/test/kotlin/blackjack/domain/card/DeckTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
